+++ src/Models/Schedule.php
@@ -65,7 +65,7 @@ class Schedule extends Model
     /**
      * Get duration in minutes.
      */
-    public function getDurationMinutesAttribute(): int
+    public function getDurationMinutesAttribute(): float
     {
         return $this->start_datetime->diffInMinutes($this->end_datetime);
     }
